Use int8_t for small integer indexing

+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/sparse_lu_solver.hpp
@@ -176,10 +176,11 @@ template <rk2_tensor Matrix> class DenseLUFactor {
     {
         constexpr bool forward_traversal = (side == TriangularSolveSide::left) == (factor == TriangularFactor::lower);
 
-        for (int step = 0; step < size; ++step) {
-            int const index = forward_traversal ? step : size - 1 - step;
-            for (int previous_step = 0; previous_step < step; ++previous_step) {
-                int const previous_index = forward_traversal ? previous_step : size - 1 - previous_step;
+        for (int8_t step = 0; step < size; ++step) {
+            int8_t const index = static_cast<int8_t>(forward_traversal ? step : size - 1 - step);
+            for (int8_t previous_step = 0; previous_step < step; ++previous_step) {
+                int8_t const previous_index =
+                    static_cast<int8_t>(forward_traversal ? previous_step : size - 1 - previous_step);
                 if constexpr (side == TriangularSolveSide::left) {
                     rhs.row(index) -= lu_matrix(index, previous_index) * rhs.row(previous_index);
                 } else {
